Electoral_process_management/requirements.txt
flask
flask_sqlalchemy
pymysql
cryptography
flask_migrate
sqlalchemy_utils
flask_jwt_extended
python-dateutil
sqlalchemy
redis